Merged new blas stuff

+++ linbox/blackbox/ntl-hankel.inl
@@ -84,12 +84,12 @@ namespace LinBox
 			       const std::vector<typename Field::Element>&v) : Father_t(F)
 	{
 		// Assumes that the input is a vector of ZZ_p else things will FAIL
-		if ( isEven(v.size()) )
+		if ( isEven(uint64_t(v.size())) )
 		{
 			std::cout << "There must be an ODD number of entries in the input vector " <<
 			"The length given is " << v.size();
 		}
-		assert( isOdd(v.size()) );
+		assert( isOdd(uint64_t(v.size())) );
 
 		this->rowDim = (1+v.size())/2; // The vector is 0..2n-2;
 		this->colDim = (1+v.size())/2;
@@ -119,12 +119,12 @@ namespace LinBox
 	       	Father_t(v.field())
 	{
 		//! @warning Assumes that the input is a vector of ZZ_p else things will FAIL
-		if ( isEven( v.size() ) )
+		if ( isEven(uint64_t( v.size() )) )
 		{
 			std::cout << "There must be an ODD number of entries in the input vector " <<
 			"The length given is " << v.size();
 		}
-		assert(  isOdd(v.size())  );
+		assert(  isOdd(uint64_t(v.size()))  );
 
 		this->rowDim = (1+v.size())/2; // The vector is 0..2n-2;
 		this->colDim = (1+v.size())/2;
